Added a default auth heder for call creation
So, fun fact. Scheduler-for-PCF (and, therefore, ocf-scheduler)
requires that every `Call` object have an authentication header
string provided for creation.

The pcf plugin, however, doesn't give you a way to specify this,
so it must be getting it from the environment or some such.

For the moment, just for the sake of testing the plugin itself,
we are setting this to "default" and marking for tomorrow guy
to figure out.