Added git-tag based release tasks to makefile
This adds a `make release` target that does a distclean, a
distbuild, and generates linux and macos binaries in
builds/ocf-scheduler-cf-plugin-vX.Y.Z, where the version is
based on the most recent git tag.

This is really meant to only be used when we cut a release to the
main branch, and it should also be done from the main branch most
likely.